+++ src/Controller/FonctionController.php
@@ -73,16 +73,16 @@ class FonctionController extends AbstractController
     {
         $data = $request->toArray();
         if (isset($data['force']) && $data['force'] === true) {
+            if (!$this->isGranted("ROLE_ADMIN")) {
+                return new JsonResponse(["error" => "Hanhanhaaaaan vous n'avez pas dit le mot magiiiiqueeuuuuuh"], JsonResponse::HTTP_FORBIDDEN);
+            }
             $entityManager->remove($fonction);
         } else {
             $fonction->setStatus("off");
             $entityManager->persist($fonction);
         }
-
         $entityManager->flush();
-
         $cache->invalidateTags(["fonction"]);
-
         return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
     }
 }
