+++ src/Controller/InfoTypeController.php
@@ -73,23 +73,18 @@ class InfoTypeController extends AbstractController
     }
     #[Route(path: "/{id}", name: 'api_infoType_delete', methods: ["DELETE"])]
     public function delete( TagAwareCacheInterface $cache,InfoType $infoType,UrlGeneratorInterface $urlGenerator, Request $request,SerializerInterface $serializer, EntityManagerInterface $entityManager): JsonResponse
-   
-    
     {
         $data = $request->toArray();
-    
+
         if (isset($data['force']) && $data['force'] === true) {
+            if (!$this->isGranted("ROLE_ADMIN")) {
+                return new JsonResponse(["error" => "Hanhanhaaaaan vous n'avez pas dit le mot magiiiiqueeuuuuuh"], JsonResponse::HTTP_FORBIDDEN);
+            }
             $entityManager->remove($infoType);
-            
-
         } else {
-            $infoType
-                ->setStatus("off")
-            ;
-
+            $infoType->setStatus("off");
             $entityManager->persist($infoType);
         }
-
         $entityManager->flush();
         $cache->invalidateTags(["infoType"]);
         return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
